requirements.txt
requests==2.20.1
numpy==1.15.2
scipy==1.1.0
scikit-learn==0.20.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cloudpickle==0.5.6
cycler==0.10.0
idna==2.7
kiwisolver==1.0.1
matplotlib==3.0.1
pandas==0.23.4
Pillow==5.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
torch==1.13.1
torchvision==0.2.1
tqdm==4.28.1
ujson==1.35
urllib3==1.24.1
